feat(router): add AButtonView and move all code from App.vue
